Rework how chronicle_kv's ets table is published.
Instead of using the rename trick to achieve atomicity (which can't be
extended very easily to be used with rsm snapshots). With the change
chronicle_kv maintains two tables: one for meta-information and one
for kv pairs. The meta-information table holds a reference to the
kv-table. When installing a snapshot, a new table will be created and
populated and then atomically swapped with the old one by updating the
reference.