Always truncate diverged entries before logging metadata.
It's important to truncate diverged entries before logging
metadata. We only truncate on the first append in a new term. So the
metadata will include the new term_voted value. But if truncate
happens after this metadata gets logged, if the process crashes, upon
recovery we might end up with an untruncated log and the new
term_voted. Which would be in violation of the following invariant:
all nodes with the same term_voted agree on the longest common prefix
of their histories.