Prevent removed node from getting votes from half-added nodes.
When a node is being added to a cluster, it starts in an empty
state. So it'll accept vote solicitations from any node with the
correct history id. But it's possible that the node will get contacted
by nodes that were previsouly removed from the cluster (if those don't
know that they got removed). For this to happen the node being added
must have previsouly been part of the cluster together with the
removed nodes. The latter nodes will have the correct history id and
will be able to get the added node to vote for them. Which may give
one of those nodes enough votes to become the leader, which must never
happen.

To protect against this the node will reject vote solicitations from
nodes that are not aware of the latest config as of when the node add
sequence was initiated.