Don't truncate entries after the snapshot when they may be valid.
When handling install_snapshot(). Specifically in the case where
CommittedSeqno < SnapshotSeqno =< HighSeqno.

Install snapshot always tosses away all entries in the log and
replaces them with the snapshot. In this case, some or all of the
entries following the snapshot might ultimately be committed and must
be preserved. Otherwise in certain situations it's possible for a
committed entry to get overwritten.

It's only possible for the tail of the log after the snapshot
to have "valid" entries if the entry at SnapshotSeqno is
valid. We check if that's indeed the case by comparing the
entry's term and the snapshot term. If they match, then there's
no need to install the snapshot. We can simply mark everything
up to the snapshot seqno committed.