Don't send history id with commands.
With recent changes commands from partitioned off nodes won't get
applied because only command from current peers get applied.

When it comes to get_quorum_revision, it's possible that we'll respond
to a partitioned out node. But the response will include the new
history id, so this will be caught once the client tries to wait for
the revision.