Deal with leaders competing for the same term slightly better.
Previously it was possible for two (or more) leaders trying to
establish the same term to split votes in half and then get stuck
waiting for a remaining vote (which may not arrive due to the node
being down or slow). This patch lowers the timeout significantly when
such situation is detected.