Don't terminate chronicle_server when proposer crashes.
This used to be needed because some of the requests may not have been
given a response, and so the only way for clients to get unblocked was
for chronicle_server (which clients would monitor) to terminate. With
proper retry logic in chronicle_rsm, there's no need to terminate
chronicle_server anymore. But some care is needed to make sure that
requests don't get multiple responses.