Handle truncating appends synchronously.
When an append needs to truncate an uncommitted tail of the log, it
needs to overwrite some of those uncommitted entries in memory, so
they can be used by the subsequent appends for validation (even before
the truncating append is fully persisted). But that makes the
in-memory log as of the persisted high seqno inconsistent. So this
state may not be exposed. This patch makes truncating appends
synchronous, so outside consumers can either see the state before the
truncating append, or after it, but cannot see the intermediate state
where the truncating append is not yet persisted and the in-memory log
is inconsistent.