Add function to return default txn timeout
This adds the get_txn_default_timeout/0 function which returns the
default timeout for a chronicle transaction. This can be used by clients
of chronicle transactions to determine higher level timeouts for multi
step operations, of which a chronicle transaction is one step.

Even though chronicle transactions support specifying a timeout, knowing
the default value allows clients to calculate an upper level timeout
without having to know what the specific chronicle default is.