MB-56910 Make chronicle more resilient to corrupted sealed logs
The current chronicle will crash on reading the log files if there's
an inconsistency between sealed logs and the current one, despite the
fact that sealed logs are completely irrelevant for the further
chronicle lifecycle.

After this change, chronicle will print an error and then will try
to read just current log file in isolation.

Change-Id: I6b3feb54d8fa36a7f38db79f2cc19e1ea31c6ea4