MB-60458: Fix dialyzer warnings in chronicle_storage
In chronicle_storage we have dialyzer warnings that we cannot match a
couple of clauses as the function chronicle_utils:sync_dir/1 should
always return ok. One case clause can be removed, but the other is used
in a generic macros which wraps various other function calls. The macro
could be copied without the unneeded case, but the code re-use here
seems acceptable. We could also turn the macro into a function, it's not
clear if a macro is/was even necessary or appropriate here, but it's
much less invasive to turn off the warning for this function.