MB-63122: Add artifacts declaration for chronicle_dump
rebar3 is not capable of determining that this file needs to be rebuilt
if the chronicle erlang code does not need to be rebuilt. The Server
build requires that it exists, it's necessary for log collection, so we
need to ensure that it is built.

We need to do this for priv/sync_nif.so too, but that's somewhat
troublesome as we need to do that only for linux. We'll address that in
a future change.